% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IC_based_SE.R
\name{IC_based_se}
\alias{IC_based_se}
\title{Calculate the influence curve based SE with given fitted HAL object and evaluation points}
\usage{
IC_based_se(
  X,
  Y,
  hal_fit,
  eval_points,
  family = "binomial",
  X_unpenalized = NULL
)
}
\arguments{
\item{X}{An input \code{matrix} with dimensions number of observations -by-
number of covariates that will be used to derive the design matrix of basis
functions.}

\item{Y}{A \code{numeric} vector of observations of the outcome variable.}

\item{hal_fit}{The HAL fit object.}

\item{family}{A \code{character} or a \code{\link[stats]{family}} object
(supported by \code{\link[glmnet]{glmnet}}) specifying the error/link
family for a generalized linear model. See more details in \code{fit_hal}.}

\item{X_unpenalized}{TAn input \code{matrix} with dimensions number of observations -by-
number of covariates that are not penalized when fit HAL.}
}
\description{
Calculate the influence curve based SE with given fitted HAL object and evaluation points
}
\details{
The procedure calculates the SE with the following steps:
    1). get all the selected basis functions and corresponding from hal_fit 
    2). calculate influence curves for coefficients 
    3). use delta method to calculate influence curves for points on the curve
}
