% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_DRC.R
\name{fit_UHAL_DRC}
\alias{fit_UHAL_DRC}
\title{fitting HAL and learn the dose response curve}
\usage{
fit_UHAL_DRC(
  dat,
  y_var_name,
  trt_var_name,
  family,
  curvePoints = NULL,
  Unsersmoothing = TRUE,
  smoothOrderAdapt = FALSE,
  smoothOrder = 1,
  baseNumKnots = NA,
  boundResults = TRUE
)
}
\arguments{
\item{y_var_name}{A \code{character} of the outcome variable name}

\item{trt_var_name}{A \code{character} of the treatment variable name}

\item{family}{A \code{character} or a \code{\link[stats]{family}} object
(supported by \code{\link[glmnet]{glmnet}}) specifying the error/link
family for a generalized linear model. See more details in \code{fit_hal}.}

\item{curvePoints}{A \code{numeric vector} with points on the dose response
curve that wants to be estimated.}

\item{smoothOrderAdapt}{A \code{boolean} indicating whether data-adaptivly
selects the smoothness order and base number knots for HAL fitting.}

\item{smoothOrder}{A \code{numeric} smoothness order number of HAL fit.}

\item{baseNumKnots}{A \code{numeric} base_num_knots of HAL fit.}

\item{boundResults}{A \code{boolean} indicating whether bound the predicted
values and confidence intervals by the minimum and maximum of the provided Y.}

\item{data}{An input \code{data.frame} with dimensions number of observations -by-
number of covariates plus the outcome variable.}
}
\description{
fitting HAL and learn the dose response curve
}
\details{
The procedure fits undersmoothed HAL with provided data and HAL fitting
hyperparameters, then it learns predicts dose response curve using the fitted
HAL as a working model and provide delta-method based CI.
}
